Renames test local --namespace flag --operator-namepsace
This patches chages the `--namespace` flag in `operator-sdk test local <test-dir> --up-local`
command with `--operator-namespace`. This is done with respect to
the discussion (git and slack) #2100 #2149

This will help with making the behavior and scope of the flag explicit.
Now a flag `--watch-namespace` flag can be added to `operator-sdk test local <test-dir> --up-local`
to explicitly set `WATCH_NAMESPACE` envvar to different namespace (even `WATCH_NAMESPACE=""`)

Signed-off-by: Nikhil Thomas <nikthoma@redhat.com>